Code from my Spaceships game
Hello this is part of my latest Unity project:

the following scripts were taken out of a bigger project of mine that i cannot excess right now because of a misunderstanding regarding my unity account. I am currently seeking active support form unity support! 

the game I've built is an old school spaceship up and down game which is built on a tier system - there are 10 levels and when the player finishes the 10 levels he will advance to tier 2 which is the same 10 levels but with different sprites and different stats according to the current tier, game ends when the player fails or finish tier 7 (70 levels) 

The shopUI script is about the shop UI of course! which i implemented there are 2 shield upgrades, 3 weapons & 3 projectiles & 3 Engine upgrades respectfully and the option to heal and buy armor. also added a fun little description to each upgrade and a cost that is manageable within the respectful levels.

The spawnManager script is about the spawning of the enemies of the current level and keep track whenever the level has ended. 

The folder holds 12 different scripts i built for different enemies and 2 of them are bosses and one of them even has a second phase each with his unique movement and effect / projectile